Move static method
