+++ request-response/logbook/src/main/java/no/entur/logging/cloud/logbook/AbstractLogLevelLogstashLogbackSink.java
@@ -15,30 +15,6 @@ import java.util.function.BooleanSupplier;
 
 public abstract class AbstractLogLevelLogstashLogbackSink extends AbstractLogLevelSink {
 
-    public static boolean isXmlMediaType(@Nullable final String contentType) {
-        if (contentType == null) {
-            return false;
-        }
-
-        String contentTypeWithoutEncoding;
-        // text/xml;charset=UTF-8
-        int index = contentType.indexOf(';');
-        if(index == -1) {
-            contentTypeWithoutEncoding = contentType;
-        } else {
-            contentTypeWithoutEncoding = contentType.substring(0, index).trim();
-        }
-
-        final String lowerCasedContentType = contentTypeWithoutEncoding.toLowerCase();
-
-        boolean isApplicationOrText = lowerCasedContentType.startsWith("application/") || lowerCasedContentType.startsWith("text/");
-        if(!isApplicationOrText) {
-            return false;
-        }
-
-        return lowerCasedContentType.endsWith("+xml") || lowerCasedContentType.endsWith("/xml");
-    }
-
     protected final MaxSizeJsonFilter maxSizeJsonFilter;
     protected final JsonValidator jsonValidator;
