+++ request-response/logbook/src/main/java/no/entur/logging/cloud/logbook/AbstractLogLevelSink.java
@@ -3,11 +3,36 @@ package no.entur.logging.cloud.logbook;
 import org.slf4j.Marker;
 import org.zalando.logbook.*;
 
+import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.function.BiConsumer;
 import java.util.function.BooleanSupplier;
 
 public abstract class AbstractLogLevelSink implements Sink {
+
+    public static boolean isXmlMediaType(@Nullable final String contentType) {
+        if (contentType == null) {
+            return false;
+        }
+
+        String contentTypeWithoutEncoding;
+        // text/xml;charset=UTF-8
+        int index = contentType.indexOf(';');
+        if(index == -1) {
+            contentTypeWithoutEncoding = contentType;
+        } else {
+            contentTypeWithoutEncoding = contentType.substring(0, index).trim();
+        }
+
+        final String lowerCasedContentType = contentTypeWithoutEncoding.toLowerCase();
+
+        boolean isApplicationOrText = lowerCasedContentType.startsWith("application/") || lowerCasedContentType.startsWith("text/");
+        if(!isApplicationOrText) {
+            return false;
+        }
+
+        return lowerCasedContentType.endsWith("+xml") || lowerCasedContentType.endsWith("/xml");
+    }
 
     protected final BooleanSupplier logLevelEnabled;
